Add test case which shows failure to combine frames together
